% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{detrend}
\alias{detrend}
\title{De-trends a numeric vector}
\usage{
detrend(x, window_size = 5)
}
\arguments{
\item{x}{A numeric vector.}

\item{window_size}{A length for the rolling window to calculate the median.}
}
\value{
A numeric vector, de-trended.
}
\description{
Given a numeric vector, it calculates a rolling median and then from each element, it subtracts the local median. The window for the median is centered on the element, so if the size is 5, it takes the element itself and the elements up to ±2 positions from it. If the size is even, it's rounded down (e.g. window_size = 6 is the same as window_size = 5).
}
