% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{identify_milestones}
\alias{identify_milestones}
\title{Identify milestone years and papers}
\usage{
identify_milestones(
  CR_data,
  n_milestones = 10,
  n_refs = 1,
  min_times_cited = 2,
  cited_year_range = NULL
)
}
\arguments{
\item{CR_data}{A list containing the bibliometric data frame, as returned by *run_CRE()*}

\item{n_milestones}{How many milestone years will be selected (default is 10).}

\item{n_refs}{How many references from each year will be selected (default is 1).}

\item{min_times_cited}{References cited strictly less than this argument will be excluded from the analysis (default is 2)}

\item{cited_year_range}{Passed to the plotting function (the analysis will take up the full range)}
}
\value{
A list containing a data frame with the milestone papers, a vector of the milestone years and two options of RPYS figures (ggplot objects)
}
\description{
Given a data frame of cited references, it runs a multi-RPYS analysis
to find the milestones. Details of the algorithm can be found in the
following references:
}
\details{
Comins & Hussey (2015). Compressing multiple scales of impact detection by Reference Publication Year Spectroscopy. Journal of Informetrics. DOI:10.1016/j.joi.2015.03.003
Comins & Leydesdorff (2017). Citation algorithms for identifying research milestones driving biomedical innovation. Scientometrics. DOI:10.1007/s11192-016-2238-1
}
