% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{plot_multi_rpys}
\alias{plot_multi_rpys}
\title{Plots a multi-RPYS plot}
\usage{
plot_multi_rpys(CR_data, cited_year_range = NULL)
}
\arguments{
\item{CR_data}{A list containing the bibliometric data frame, as returned by *run_CRE()*}

\item{cited_year_range}{The range of years to plot. Default is NULL, which the function transforms to "from 1900 to the most recent year in the dataset".}
}
\value{
A ggplot.
}
\description{
The plot shows a matrix where the rows indicate the citing year and columns the cited year. The color indicates the frequency of the cited year for that citing year (i.e. colors are relative to each row). However, to be comparable across citing years, the color indicates the ranking, not the absolute values of the deviations from the median of the 5-years around it.
}
\details{
In other words, a simple RPYS analysis is done for each citing year, the de-trended data is calculated (i.e. deviation from the local median) and the de-trended data is then ranked. This is the same procedure followed for identifying milestones.
}
