% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{plot_rpys}
\alias{plot_rpys}
\title{Plots a single RPYS plot}
\usage{
plot_rpys(CR_data, cited_year_range = NULL)
}
\arguments{
\item{CR_data}{A list containing the bibliometric data frame, as returned by *run_CRE()*}

\item{cited_year_range}{The range of years to plot. Default is NULL, which the function transforms to "from 1900 to the most recent year in the dataset".}
}
\value{
A ggplot.
}
\description{
The plot shows the frequency of cited references (CR) in each year, as well as
the detrended value (#CR - median of the #CR of 5 years around (±2)).
}
