% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{run_CRE}
\alias{run_CRE}
\title{Process and disambiguate references using CRExplorer}
\usage{
run_CRE(
  data_files = NULL,
  data_path = NULL,
  cre_path = NULL,
  type = "WOS",
  cited_year_range = c(0, 2100),
  include_cited_wo_year = T,
  citing_year_range = c(0, 2100),
  include_citing_wo_year = T,
  sampling = "None",
  offset = 3,
  import_max = 10000,
  save_files = NULL
)
}
\arguments{
\item{data_files}{A character vector with a list of filenames. If provided, data_path must be NULL.}

\item{data_path}{A character vector with a directory. If provided, data_files must be NULL.}

\item{cre_path}{Path to the CitedReferencesExplorerScript.jar file.}

\item{type}{Type of bibliometric records, can be "WOS", "Scopus" or "CrossRef". Defaults to "WOS".}

\item{cited_year_range}{Vector with minimum and maximum publication year for cited references to be included. Defaults to c(0,2100).}

\item{include_cited_wo_year}{Logical. Whether to include cited references without publication year. Defaults to c(0,2100).}

\item{citing_year_range}{Vector with minimum and maximum publication year for citing references to be included.}

\item{include_citing_wo_year}{Logical. Whether to include citing references without publication year.}

\item{sampling}{Type of sampling (useful when there are many references). Can be "Random" or "None" or "Systematic" or "Cluster". Defaults to "None" (i.e. use all references).}

\item{offset}{Used only when sampling is "Systematic".}

\item{import_max}{Maximum number of references to import. Defaults to 10,000.}

\item{save_files}{If NULL (default), the data will not be saved to disk. If you provide a path, it will save the results to the given folder.}
}
\value{
A list containing two data frames, one with the references (citing and cited) and one with the data underlying the RPYS graph
}
\description{
*data_files* or *data_path* serve to specify the files to be read and
*save_files* indicate whether to save the output to disk. The rest of
the arguments are passed to CRExplorer.
}
\details{
For more details about how the arguments are used, see the documentation
for CRExplorer at https://andreas-thor.github.io/cre/manual.pdf
}
